+++ autofit/non_linear/optimize/drawer/plotter.py
@@ -1,0 +1,5 @@
+from autofit.plot.samples_plotters import MCMCPlotter
+
+class DrawerPlotter(MCMCPlotter):
+
+    pass
